Added Different Tests and Survey
Users now answer three questions before starting. In addition, the
system checks if the browser supports speech recognition, and has three
different tests configured. Next step is to assign these to users and
alter the database model so that the type of test is saved for the user
and the time it took them.